+++ src/components/crds-shared-header/global-nav/readme.md
@@ -28,15 +28,14 @@
 
 ### Depends on
 
-- [profile-nav](../profile-nav)
-- [give-nav](../profile-nav)
+- [profile-nav](..\profile-nav)
+- [give-nav](..\profile-nav)
 
 ### Graph
 ```mermaid
 graph TD;
   global-nav --> profile-nav
   global-nav --> give-nav
-  profile-nav --> nav-link
   crds-shared-header --> global-nav
   style global-nav fill:#f9f,stroke:#333,stroke-width:4px
 ```
