+++ src/components/crds-shared-header/nav-link/readme.md
@@ -7,31 +7,12 @@
 
 ## Properties
 
-| Property       | Attribute       | Description | Type     | Default     |
-| -------------- | --------------- | ----------- | -------- | ----------- |
-| `automationId` | `automation-id` |             | `string` | `undefined` |
-| `href`         | `href`          |             | `string` | `undefined` |
+| Property        | Attribute       | Description | Type       | Default     |
+| --------------- | --------------- | ----------- | ---------- | ----------- |
+| `automationId`  | `automation-id` |             | `string`   | `undefined` |
+| `handleSignOut` | --              |             | `Function` | `undefined` |
+| `href`          | `href`          |             | `string`   | `undefined` |
 
-
-## Events
-
-| Event            | Description | Type               |
-| ---------------- | ----------- | ------------------ |
-| `signOutClicked` |             | `CustomEvent<any>` |
-
-
-## Dependencies
-
-### Used by
-
- - [profile-nav](../profile-nav)
-
-### Graph
-```mermaid
-graph TD;
-  profile-nav --> nav-link
-  style nav-link fill:#f9f,stroke:#333,stroke-width:4px
-```
 
 ----------------------------------------------
