+++ src/components/crds-shared-header/profile-nav/readme.md
@@ -9,10 +9,9 @@
 
 | Property              | Attribute                | Description | Type       | Default     |
 | --------------------- | ------------------------ | ----------- | ---------- | ----------- |
-| `config`              | `config`                 |             | `any`      | `undefined` |
 | `currentUser`         | `current-user`           |             | `any`      | `undefined` |
 | `data`                | --                       |             | `JSON`     | `undefined` |
-| `onSignOut`           | --                       |             | `Function` | `undefined` |
+| `handleSignOut`       | --                       |             | `Function` | `undefined` |
 | `profileNavIsShowing` | `profile-nav-is-showing` |             | `boolean`  | `true`      |
 
 
@@ -20,16 +19,11 @@
 
 ### Used by
 
- - [global-nav](../global-nav)
-
-### Depends on
-
-- [nav-link](../nav-link)
+ - [global-nav](..\global-nav)
 
 ### Graph
 ```mermaid
 graph TD;
-  profile-nav --> nav-link
   global-nav --> profile-nav
   style profile-nav fill:#f9f,stroke:#333,stroke-width:4px
 ```
